+++ NEWS.md
@@ -1,5 +1,7 @@
 # ellmer (development version)
 
+* `chat_anthropic()` now defaults to Claude Sonnet 4. 
+* Add pricing information for latest generation of Claude models.
 * You can now use pre-existing JSON schemas in structured chats using `type_from_schema()` (#133, @hafen)
 * `chat_databricks()` now picks up on Databricks workspace URLs set in the
   configuration file, which should improve compatibility with the Databricks CLI